Add files via upload
Our goal in this project was to create a tool that allows finding type II TA proteins in a bacterial genome and then using it for the analysis of bacterial genomes and learning about their TA. In order to produce it, we developed an algorithm that uses prokka (software), which allowed us to find the proteins present in the bacterium's genome. After that, we performed a blastp run at once on all the proteins found by prokka against the type II TA protein database to find the TA proteins and isolate them from the other proteins present in the bacterium. We allowed the user to choose a cutoff for the evalue of the results, in order to filter out results that are not close enough to the queries of the proteins that prokka found and are probably unlikely to be indeed proteins that exist in bacteria. We performed additional filters such as downloading results that matched the same query, and leaving the result that was most similar to the query. After the filters, we got the list of toxins and antitoxins suspected to be in the bacterial genome.
Later in the project, we downloaded genomes of strains of Pseudomonas aeruginosa and using the tool, we checked which TAs are in the genomes and built a graph of the total total proteins we found for each strain. Next, we wrote code that performs various analyzes on the tool files. We built graphs describing the TA proteins common to the genomes of the strains on which we performed the analysis, and the TA proteins unique to each strain. In addition, we checked the distances between the TA locations found in order to check the presence of island TA in the examined genomes and built a graph depicting the number of island TA found in each strain.